requirements.txt
Authlib==0.11
beautiful-date==2.0.0
Brotli==1.0.9
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
charset-normalizer==2.0.3
chart-studio==1.1.0
click==8.0.1
colorama==0.4.4
cryptography==3.4.7
dash==1.20.0
dash-auth==1.4.1
dash-core-components==1.16.0
dash-google-oauth==1.2
dash-html-components==1.1.3
dash-renderer==1.9.1
dash-table==4.11.3
discovery==0.0.4
et-xmlfile==1.1.0
Flask==2.0.1
Flask-Compress==1.10.1
Flask-SeaSurf==0.3.0
future==0.18.2
gcsa==1.2.0
google-api-core==1.30.0
google-api-python-client==2.11.0
google-auth==1.32.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
googleapis-common-protos==1.53.0
greenlet==1.1.0
gspread==3.7.0
gunicorn==20.1.0
httplib2==0.19.1
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
numpy==1.21.0
oauthlib==3.1.1
openpyxl==3.0.7
packaging==20.9
pandas==1.2.5
plotly==5.1.0
protobuf==3.17.3
psycopg2==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyOpenSSL==20.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.18.0
python-google-calendar-api==0.5
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.7.2
six==1.16.0
SQLAlchemy==1.4.20
tenacity==7.0.0
tzlocal==2.1
ua-parser==0.10.0
uritemplate==3.0.1
urllib3==1.26.6
Werkzeug==2.0.1
